Restrict bsearch block type (#7732)
Currently `Indexable#bsearch`, `Indexable#bsearch_index` and
`Range#bsearch` all take a block that accepts anything as a return
value. This is a bit unintuitive because the block should only return
`true` or `false`. It also happens that Ruby provides the same
functionality but an integer can be returned from the block and this
might confuse people coming from Ruby.