Raise SyntaxError rather than ArgumentError for 9999999...i32
That is, very large integers *with* the suffix specified.
It was unconditionally converting the value to uint64, but have to check for the validity of that first.
Additional check for num_size < 20 fixes it. Such numbers definitely fit into uint64. And this condition fully covers int64, too. gen_check_uint_fits_in_size is not used for uint64 so this is fine.