HTTP::Headers: optimize representation
Instead of storing headers internally as `Hash(String, Array(String))`
we store them as `Hash(String, String | Array(String))`. This involves
a bit more logic when dealing with this union but it saves a fair
amount of memory because for headers with just a single value, which
is the most common case, we avoid allocating memory for an array.