HTTP::Request: optimize request line parsing
Instead of using `String#split`, which creates an array with three
strings, we find the space indexes and create subslices/substrings for
each of the pieces.
We also avoid allocating a string for common HTTP methods (GET, POST,
etc.) and for the supported HTTP versions.
Finally, we use `IO#peek` to see if we can find the request line there
instead of allocating a String for it.