Compiler: don't skip all yield nodes if one is missing an exp type (#8101)
It seems when some yield expression didn't have a type we would break
out from the yield typing logic. It makes sense to skip that yield but
not all yields. This essentially changes a `return` for a `next` in a
nested loop, but because the logic was getting a bit large and confusing
a method was extracted for the contents of one loop.