Move Markdown to inside the compiler.
The Markdown module is buggy and not feature complete.
We shouldn't expose it in the standard library. For the compiler's
purposes (`crystal doc`) it might be good but users should be encouraged
to use a better library. Eventually the `crystal doc` tool should
somehow use a different, better markdown library.