Add `Hash#store`
This lets you store a key-value pair and know whether an old value was
present and what is it. We might be tempted to use a method that returns
the value or `nil` if it wasn't already there but this doesn't work if
values allow `nil` as a valid value. So `store` is in a way the
equivalent of `fetch` but when storing a value.