Delay releasing of fiber stack in multithread mode
Once the stack is released it can be used by any other fiber. This patch
ensures the release is done only after the current fiber has finished executing.

In single thread mode this change is not needed because there cannot be new fibers
until the context is switched to another fiber.