[Compiler] Add documentation for pseudo methods (#8327)
This adds documentation for pseudo-methods `!`, `as`, `as?`, `is_a?`, `nil?` and `responds_to?`, all on `Object`.
These methods are implemented by the compiler and cannot be redefined, so we can't simply document them in code. Instead, the method names are prefixed by `__crystal_pseudo_` so they can be documented like regular methods. The prefix is then stripped off by the doc generator.