Compiler: correctly erase top-level local vars after block
You might know this or not, but macros have access to local variables of
the scope where they are called. For example:

```crystal
macro id(obj)
  {{obj}}
end

a = 1
puts id(obj)
```

The above macro expands to `a`. If that's seen as a program by itself,
`a` parses to a call. However, because the macro is invoked in a context
where `a` is a local variable, `a` is parsed as a local var.

For this to work when expanding macros in the top-level phase, the
compiler keeps track of existing local variables: clearing them when
entering a new `def`, etc. A similar logic should be done for call
blocks: after a block is analyzed all local vars declared inside it
should not be visible outside it. Except that this wasn't the case: the
compiler just cleared block args after a block, but not all variables.
That means that a macro evaluated at one point in the program will have
access to each and every local variable defined before it, even if it's
not actually visible. This isn't that terrible, but in some cases it can
lead to a lot of extra logic. Specifically, a Set with the local vars is
created on each macro exapnsion and in a macro loop that keeps creating
local vars inside blocks it can lead to huge sets.