IO: Detect closed IO when resuming readable/writable event
If the IO was closed by other thread while waiting we need to check if it's open or not.

If it's closed an IO::Error is raised as a result of the read/write operation as if the IO was has been closed from the beginning.

The event_close is now performed before actually closing the underlying file descriptor to prevent libevent accessing it in a closed state.

The @closed = true in IO::FileDescriptor needs to happen before the system_close so it is available when the awaiting fibers are woken up.