Codegen: Use llvm memset and memcpy matching arch (#8746)
On 64bits llvm.memset.p0i8.i64 will be used
On 32bits llvm.memset.p0i8.i32 will be used
Same for memcpy

There is no spec in llvm that guarantees the 32bits version works in 64 bits. We should always use the right one.

Some initial regression tests using FileCheck are added that are not automated in the CI but helps as a starting point.