Fix YAML specs
Some of this meant ensuring we add the YAML::Any wrapper around the hash
keys. Honestly, it kinda made me wish we'd stuck with the recursive type
alias.

It's notable that JSON didn't have these issues because JSON is strictly
string keys, so type checks on the keys didn't break it in any way.