Correctly support websocket close codes
The first two bytes of the message in a WebSocket Close frame is a
numeric code. Make sure to always encode this correctly, to avoid being
in violation of the websocket spec. Also automatically decode this for
clients. This is a breaking change since on_close now takes two
arguments: close code and message, but this change was unable to be made
in a backwards compatible way.

Closes #6469
Closes #7483