Move verbose to a lower level than debug
Typically "verbose" is the most verbose output you can get and implies
verbose will give you "more"

Right now "debug" is lower than verbose which does not seem like it
matches how most people think of or are using these levels.

My proposal is to make "verbose" the most "verbose" log level possible.

For example and ORM might use "debug" to show what queries run, and what
the args are, but might have "verbose" mode that tells you when the
transactions start, when it was committed, whether there was a cache
hit, etc. Stuff that normally is not helpful, but if you *really* need
it you can go "verbose" and get all the nitty gritty details