Parser: rewrite type parser
For some reasons, the parser for types and generic type path literals is
known as one of the most complex part of the parser.
(Another one is `do ... end` block parsing.)

- **too much sharing**: the parser for a single type, type arguments,
  proc type arguments, tuple element types and others are unified to just one method.
  This code-sharing is excessive. They are not share syntax really, so there are many
  conditions and branches, thus it is hard to follow code flow.
- **historical naming**: a type path (or part ot this) is called as `ident` for historical reason.
  However `ident` means different today.

As the above reasons, gazillion bugs live in there unfortunately.
For instance, they are valid type restriction for now:

  - `*Foo` (orphan splat)
  - `(Foo, Bar) | Baz` (comma + union type)
  - `{(Foo, Bar), Baz}` (comma + tuple type)

This commit refines type grammer and rewrite parsers.