Resolve the always-present Windows linker's warnings
Currently when building on Windows, the version of MSVC linker is always printed first, then the list of all inputs, then a warning for every object file because it ends with .o, not .obj.

So, use the .obj extension and disable the default outputs. So errors are still printed, but normally the output is empty.