Format: avoid formatter crush on string interpolations in heredoc
Fixed #9362

Introduce StringInterpolationContent AST node to disinguish between usual
string literal and part of string interpolation content.
StringInterpolationContent inherits StringLiteral. It seems a bit tricky,
but changes are little in fact.

Note, this distinction is necessary for formatter surely.
It is impossible that lexer and formatter on heredoc sychronize without this.