Disable LLVM Global Isel
It's not enabled by default on x86, but on other targets it is.
At least until https://reviews.llvm.org/D80898 is released, it
doesn't like us generating a value of a zero sized type

We may need to fix doing that, but until then this workarounds
the issue.

See https://github.com/crystal-lang/crystal/issues/9297#issuecomment-636512270
for more background.