Fix error while parsing the subcommand with a hyphen
Subcommands can have a hyphen in the name. The parser treats this as
a flag and some junk value. For example, `sub-command` matches the
pattern `/-(.)\S+/` and sets flag as `su`(First two letters) and
Value Required for this flag(`FlagValue::Required`).

With this PR, Regex is modified to check if hyphen exists in the
beginning or not.

Signed-off-by: Aravinda Vishwanathapura <mail@aravindavk.in>