Fix `CRYSTAL_OPTS` parsing
Fixed #9509
Ref https://github.com/crystal-lang/crystal/pull/8900#discussion_r391925011

This commit adds `Process.split`, which works as like `Shellwords.split`
in Ruby.
And, this is used to parse `CRYSTAL_OPTS`.

I'm not sure `Process.split` is good naming. However I have no idea of
better name.
And also, I'm not sure Windows version of `Process.split` is needed.