Make Errno enum values a set union of all possible LibC Errno values
So, if an error name is available on at least one platform, it is possible to refer to it on all platforms, even if its value is `INVALID = -1`.
That is in accordance with Crystal's planned API guarantee: if it compiles on one platform, it should compile on any platform.
This is not problematic because all usages should be of the form `if err == Errno::ENOATTR`, and if such a value is not possible to encounter on a given platform, it's expected that the equality will never be true. Additionally, if the platform in the future gains more possible error codes that we're not aware of yet, the equality with them will also be false (they are not clumped into `INVALID` but rather have their normal integer value, just that it's not known to the `enum` -- `to_s` and such).

Also update and sort the list of all errors accordingly.