Add HTTP::Client logging and basic instrumentation
This adds protected methods HTTP::Client#before_exec and #after_exec that can be used to hook into the lifecycle of the http client.

A first utility for this is adding a http.client log source to know a request is starting.

This enables some basic instrumentation to monitor the activity of the application without monkey patching sensitive parts of the http client