Add OpenSSL::SSL::Context::CIPHER_SUITES_* constants
Use colons instead of spaces for ciphers

The cipher list consists of one or more cipher strings separated by colons. Commas or spaces are also acceptable separators **but colons are normally used**.
Ref: https://www.openssl.org/docs/man1.0.2/man1/ciphers.html

For ciphersuites only colons are allowed as separators, so it's better to use a separator that works for both.