Improve documentation for Object#to_s and #inspect
This patch changes the method definitions as proposed in #9966 and adds some
additional notes.

The part about ambigouity of collection members is removed because that's not
generically relevant for `#inspect` but for individual implementations such as
`Array#inspect`. So I added a short note about it there.
It does not need to be repeated for every collection type.