Implement segfault handler in Crystal
Required adding some C bindings for various signal structures,
based on rust's fairly comprehensive libc bindings.

This removes the need to build and link a separate C library and
reduces the complexity of cross-compiling (at build-time), with
some added complexity when porting to a new platform.

`siginfo_t` can be fairly complex, involving large context-dependent
unions that vary across architecture and platform. In Crystal's
case, we only care about the segfault version of the structure,
so we omit the others.