Remove Cookie Name Decoding
Crystal automatically encodes and decodes cookie names. The cookie
spec[1] encourages the value to be encoded but says nothing about
the cookie name (it could be encoded or not).

While encoding is allowed, a vunerability can occur if the cookie
name is decoded in such a way that a cookie not using a correctly
formed cookie prefix will overwrite a cookie that is using a
proper cookie prefix rendering the protections provided cookie
prefixes void.

Since Crystal is encoding the entire cookie name using URL encoding
two cookies can be sent by the browser and if in the correct order
the server will read the cookie without the proper cookie prefix.
An example cookie header value is:

    __Secure-Name=safe; __%53ecure-Name=evil

The exact same issue was recently corrected in the `rack` library
for the Ruby language[2] as well as the `werkzeug` library used
by Python's Flask web framework[3]. In fact those corrects are
what prompted me to look to see if Crystal had the same issue.

Although NIST rated the Rack vunerability "high"[4] the Rack
maintainers consider that a mistake and consider is a low priority
vunerability[5]. I agree.

The solution taken by both `rack` and `werkzeug` were to simply
stop encoding the cookie name. If an application needs it encoded
because they have a funky name that might mess with header parsing
that application can do the encoding itself. This is the solution
implemented on this commit.

An alternative solution that is probably more compatible but
also more complex is to encode the part of the cookie name after
the cookie prefix if the cookie has a prefix that matches
one of the official cookie prefixes (`__Secure-` and `__Host-`).
While this would preserve the encoding for the most part I
don't feel the complexity introduced is worth the benefit.

1. https://tools.ietf.org/html/rfc6265#section-4.1.1
2. https://hackerone.com/reports/895727
3. https://github.com/pallets/werkzeug/pull/1965
4. https://nvd.nist.gov/vuln/detail/CVE-2020-8184
5. https://github.com/rack/rack/pull/1677#issuecomment-651343672