Add return type restrictions: explicitly specify Nil
This is nice because (1) it's clear that the function is meant to be side-effect-only and (2) you don't accidentally end up returning some value.

This change affects only methods that already always return nil but didn't have the type annotation, so it should be backwards compatible.

(Sadly, there are also many places that *really should* return nil but don't. At least we won't regress on these ones.)