Retry HTTP::Client requests once if io is closed
Earlier only reads would be retried, but a very common scenario in the
wild is that the HTTP server resets the client connection after some
time. When the persistent HTTP::Client then did a request it would fail,
and it couldn't even be recovered because `close` would raise an
exception too, never setting `@io` to nil.

This change makes HTTP::Client::Request.from_io raise an IO::EOFError if the
header can't be read, instead of simply returning nil. That exception is
caught one, the `@io` closed (by ignoring IO::Error in close), and retries
the request once more, next time a proper exception is raised to the
user.

Without this patch `HTTP::Client` can't be used in any real-world
scenarios using HTTP keepalive, at least not without monkey-patching
`HTTP::Client#close` and catching and retrying your self, not very user
friendly. Also confusing when read errors are retried but not write
errors.