Add `Char#unicode_escape`
The new method is used for `#inspect` and `#dump` methods of `Char` and
`String`. It unifies the format of Unicode escape sequences to upper-cased
four-digit style (`\uFFFF`).
Char specs are updated to use the `assert_prints` helper, which
revealed a bug in `Char#dump(IO)` where all results where double quoted.
Finally, this patch adds a special rule for `'\0'.inspect` to use the
special zero escape sequence instead of Unicode escape `\u0000`.