Report syntax error for too-long bin/hex/oct integer literals
Currently the OverflowError causes a compiler crash.

Change the operations to bitwise to remove the actual overflow detection but explicitly deduce that an overflow must have happened, through the digit count.

Add the actual number literal to the error message (note: needed to ensure that the literal is fully read before declaring the overflow).