Add Enumerable#find!
Often when using the regular `find` method in `Enumerable` there is
already knowledge that the item which is looked for exists. That forces
a pattern like `array.find {..} || raise "Unreachable"`. This adds a
convenience version of find that will raise instead of returning nil if
the looked for element isn't found.