Add Enumerable#index!
Often when using the regular `index` method in `Enumerable` there is
already knowledge that the item which is looked for exists. That forces
a pattern like `array.index {..} || raise "Unreachable"`. This adds a
convenience version of index that will raise instead of returning nil if
the looked for element isn't found.