Fix `it_expands_path` spec helper with string parameters
The helper forwarded `home` and `base` paths as strings which meant that
they are converted to native paths in `Path#expand`. This is not
correct. They should be paths of the same kind as `path`.
This patch fixes that and transforms `home` and `base` paramters to
respective `Path` instances.

The same problem exists with the environment variable for the home
directory. It cannot be used, except when explicitly requested via
`env_home` variable. The change from env var to `home` parameter is
irrelevant for the specs that don't explicitly use `env_home`.