Allow explicit usage of libiconv
Allow developers to opt for libiconv variant instead of the built-in
iconv support of their platform.

This helps with cross-compilation of binaries to platforms like macOS,
which requires bulky SDK being installed in order to support linking to
their bundled `iconv` library.

Alternatively, developer should be able to build a static version of
iconv (`libiconv`) and be able to link with it by forcing its usage with
`-Duse_libiconv`:

  $ crystal build app.cr \
      --cross-compile --target x86_64-apple-darwin -Duse_libiconv

Relevant links:
- https://forum.crystal-lang.org/t/macos-compilation-and-libiconv-support/4439
- https://github.com/ziglang/zig/issues/10485#issuecomment-1013723542