Experimental: restriction augmenter (#12103)
The gist is that if you assign a method argument to an instance variable, but that method argument doesn't have a type restriction, to automatically add it. After all, if we are going to assign that value to the instance var, the type must fit in the instance variable's type.

This isn't always possible to do:
- If the method argument is assigned some other value before it's assigned to the instance var, we can't assume its type must match that of the instance var

What are the benefits? All of these!

# Better error messages

Consider this code:

```crystal
class Foo
  @x : Int32

  def initialize(x)
    @x = x
  end
end

Foo.new('a')
```

Before this PR you get this error:

```
In foo.cr:5:5

 5 | @x = x
     ^-
Error: instance variable '@x' of Foo must be Int32, not Char
```

With this PR you get this error:

```
In foo.cr:9:5

 9 | Foo.new('a')
         ^--
Error: no overload matches 'Foo.new' with type Char

Overloads are:
 - Foo.new(x : ::Int32)
 ```

# Better docs

The added types will appear in docs.

Before:

![image](https://user-images.githubusercontent.com/209371/172627295-d0a70308-e3f9-44be-abab-7d83f5af12a0.png)

After:

![image](https://user-images.githubusercontent.com/209371/172627383-50f341a7-7639-4fed-a026-211466fca99d.png)

# Autocasting works

This code:

```crystal
class Foo
  @x : Int64

  def initialize(x)
    @x = x
  end
end

Foo.new(1)
```

Before this PR it gives an error:

```
In foo.cr:5:5

 5 | @x = x
     ^-
Error: instance variable '@x' of Foo must be Int64, not Int32
```

With this PR it just compiles ®️ 

# Is this worth it?

I don't know. I think it just makes sense to do this. Maybe the exact examples above aren't common. But for example this:

```crystal
class Foo
  def initialize(@some_bool = false)
  end
end

Foo.new(1)
```

Before this PR you get this error:

```
In foo.cr:2:18

 2 | def initialize(@some_bool = false)
                    ^---------
Error: instance variable '@some_bool' of Foo must be Bool, not Int32
```

With this PR:

```
In foo.cr:6:5

 6 | Foo.new(1)
         ^--
Error: no overload matches 'Foo.new' with type Int32

Overloads are:
 - Foo.new(some_bool : ::Bool = false)
```

The error is much better, and points to the actual place where the mistake happened.

# Is there a way to disable this behavior?

Yes! If you compile your program with `-Dno_restrictions_augmenter` then this logic isn't applied. The idea is to have this flag as an escape hatch in case it's not working properly. If everything goes well we can remove the special `no_restrictions_augmenter` flag.