Support libffi on Windows (#12200)
Builds the libffi static library on Windows CI and bundles it with the build artifacts.

It uses https://github.com/winlibs/libffi, which is at version 3.3. It is not the most recent version, but that repository seems easiest to integrate into our CI. (Weirdly, the upstream provides an MSVC solution for AArch64 static builds only.) It is assumed #12192 will take care of any potential issues when upgrading to 3.4.x.