Use LibC.malloc instead of GC.malloc for LibPCRE allocations (#12456)
This makes the memory management os LibPCRE manual for the stdlib. So any internal allocations are made with the system allocator, leaving less memory being managed by the GC.

Public interface is unchanged.

As an adicional benefit: libpcre doesn't need to be linked anymore if the program doesn't create any regex!