Add more customizations to Crystal::Config.description
* Allow passing in git sha and date to Makefile
* Remove empty parens when date is ""
* Add message when compiler was not compiled in release mode

When building in a nix derivation, there is no git directory, so you
have to pass in the sha manually. Also, all the file modification
times are set to epcoh, and I think it'd be nicer to just have the
builds not have a date rather than the obviously wrong 1970-01-01.

I don't think we need to always have release mode information, but it
would be useful I think to have a way to see when you're working with a
non-release compiler.