Fix handling of quoted boolean values in `YAML::Any`
`"yes"` and `"no"` should be treated as strings (since they're quoted)
whereas `yes` and `no` should be booleans (since they're plain). This is
already handled in most places, but `YAML::Any` passes the node value,
rather than the node itself when handling a scalar and so the quoting
type is lost, and everything is parsed as though it is unquoted.