Clarify single-byte docs for `String#byte_slice`
This is a documentation-only change.

`String#byte_slice` can lead to unexpected behavior when slicing multi-
byte characters. The docs mention this, but equate "single-byte" with
"ASCII", the latter of which is a subset.

For example, `ñ` is not ASCII but is a single-byte character.

    str = "so\xF1ar"

    str.byte_slice(3) == str[3..]
    # => true

This updates the docs, effectively replacing "ASCII" with "single-byte".