MT: workaround regular timeouts in HTTP::Server specs
The core of the issue is a synchronization issue between fibers that
test edge cases in HTTP::Server, for example what's happening after
closing the fiber (from another fiber) while we have no guarantee that
the server is actually listening.

I couldn't find a proper way to fix the issue. The only solution that
fixed the reproducibility of the issue was adding a one millisecond
sleep after starting the server.

closes #13211