MT: local thread/scheduler vars in Crystal::Scheduler.enqueue
This is mostly for readability since `Thread.current` is usually a
`@[ThreadLocal]` and the result may be cached locally by LLVM to avoid
multiple resolutions, though that won't be the case for targets that
don't support ThreadLocal and for which we must use thread data storage
(e.g. Android, OpenBSD).