MT: move set_current_thread to Fiber (from Crystal::Scheduler)
This avoids manipulating `fiber.@current_thread` which ain't very
pretty, and moves the responsibility to manipulate it to
Crystal::Scheduler. That Crystal::Scheduler is also responsible from
making sure a fiber will always be enqueued or resumed on the thread
it's been associated to.

Lastly, we remove the current_thread store that was always replacing any
previous value on context swap, which sadly doesn't seem to improve
performance...