MT: move set_current_thread to Fiber (from Crystal::Scheduler)
This avoids manipulating `fiber.@current_thread` which ain't very
pretty.

Crystal::Scheduler is now entirely responsible for setting the current
thread. It's also responsible to verify that a making sure a fiber will
always be enqueued or resumed on the thread it's been associated to,
which is only affecting the `#resume(fiber)` and `#yield(fiber)` methods
that are barely used.

Lastly, we remove the current_thread store that was always replacing any
previous value on context swap. Sadly, this doesn't seem to have any
noticeable impact on performance.