Tokenize CRYSTAL_DEBUG variable
So we match exact references to 'gc' and 'sched' and expect the
delimiter to be a comma (','). This allows to remove  the `strstr(3)`
libc call that was previously introduced and the tokenization and
comparison now relies on slices and pure crystal (no need to bring
`strsep(3)` or `strtok(3)` libc calls).