Fix: opening/reading from fifo/chardev files are blocking the thread
Disk files are always opened in a blocking way because epoll (for
instance) will always say that a file is ready for reach or write. But
this is only for regular disk files, and that doesn't apply to pipes and
characters devices for example.

Opening `/dev/tty` (a character device) will block the current thread
when trying to read from it until you start typing on the keyboard.
IO::FileDescriptor was taking care to check the file type, but File
always told it to be blocking.

Worse, trying to open a fifo/pipe file will block the current thread
until another thread or process also opens it, which means that we must
determine whether to block _before_ trying to open a file (not change it
afterwards).

I also added a `blocking` parameter to `File.open` and `File.new`.