Codegen: on demand distribution to forked processes
Instead of pre-slicing the list of compilation units then having each
forked process handle its personal list, with some finishing before the
others, this patch adds bidirectional pipes to push the compilation
units as the forked processes make progress, so that each forked process
will continue to compile for as long as there is something to compile.

The forker processes are overqueued with one compilation unit to avoid
latencies (they compile the next unit while the main process pushes the
next).

This only improves performance when there are multiple compilation
units. It won't have any effect when `--single-module` (implied by
`--release`) is set, and will have little impact when a module is extra
large (e.g. crystal specs) unless maybe when there are multiple of them
and they were set to the same forked process.