Fix: Crystal::RWLock should be a struct
The Crystal::RWLock type is used to protect the GC against fiber context
switches, to avoid the GC interrupting a thread while it is switching
contexts, or starting a collection while a thread was preempted by the
kernel while it was switching the fiber context.

Problem is: Crystal::RWLock is a class, and thus depends on the GC to
be allocated, and we end up in an infinite loop: GC depends on
Crystal::RWLock that itself depends on GC, ...

The weirdest part is that it's working, and I have no idea how.