Log: delete source from builder cache when finalize
The cache of logs in Log::Builder is using a WeakRef for the Log object
to not prevent the GC from collecting them, that would leak memory, but
it keeps the hash entry, which also leaks, keeping strings (potentially
dynamically allocated) along with entries in the hash (infinitely
growing hash).

This patch adds a finalizer to Log, as suggested by @beta, to remove the
log entry from the builder cache when a Log object is collected.