Leverage Crystal::System.print_error + StringPool
Renames `print_error(fmt, *args, &)` as `printf(fmt, *args, &)` as it
involves formatting a string, not printing the formatted string to the
standard error. Also introduces `print_error(bytes)` to do the actual
print to standard error (without formatting) so we can do both format
then print to error in two steps.

Since crystal printf is a subset of C printf and we don't support
printing floats, the trace is changed to print durations as nanoseconds
integer + fix the tick to be the monotonic time since the program
started instead of the current monotonic time (should be more useful).

Last but not least, fixes the 'crystal tool trace stats' command to
parse an integer instead of float (removes dependency on `strtod(2)`)
and to leverage StringPool instead of the fixed list of static strings,
so adding or removing sections/actions won't affect the command.