Add Crystal::System::Time.ticks
It returns the current time from the monotonic clock in nanoseconds with
additional guarantees that it won't raise or allocate memory.

Note: the Win32 functions are guaranteed to never fail since Windows XP
onwards, and on posix, clock_gettime may only fail on invalid input, so
we can safely outlook errors.