OpenSSL: deprecate Mozilla TLS Server recommendation
We shouldn't hardcode a list of ciphers. It should always be a
configurable list of ciphers. Furthermore, restricting the ciphers is
only part of the security recommendations: TLS methods, curves and so on
must also be considered.

This patch deprecates the constants, and makes the related methods as
NOOP since they didn't work as expected on OpenSSL 1.1 and later, as
they only changes the cipher suites for TLS v1.3 (identical to default)
and didn't change the ciphers list for TLS v1.2 and below (oops).