Do not strip the macOS target triple (#14466)
Starting with Xcode 15, the minimum deployment target is required in the target triple.

With the release of Xcode 15, a [new linker was introduced](https://developer.apple.com/documentation/xcode-release-notes/xcode-15-release-notes#Linking) and with it came warning messages like this:

    ld: warning: no platform load command found in '/Users/pascal/.cache/crystal/Users-pascal-Documents-tutorials-crystal-hello_world.cr/F-loat32.o', assuming: macOS

It appears that the new linker is stricter in which target triples it considers valid. Specifically, it looks like the minimum deployment target is required. E.g., `aarch64-apple-darwin23.3.0` is valid, while `aarch64-apple-darwin` is not. See https://github.com/crystal-lang/crystal/issues/13846#issuecomment-2040146898 for details.

This patch removes code which strips the minimum deployment target in `LLVM.default_target_triple` as an effort for standardization.

See also: https://github.com/crystal-lang/distribution-scripts/pull/296