Add `Crystal::Tracing` for runtime tracing (#14659)
Implements tracing of the garbage collector and the scheduler as per #14618

Tracing is enabled by compiling with `-Dtracing` then individual tracing must be enabled at runtime with the `CRYSTAL_TRACE` environment variable that is a comma separated list of sections to enable, for example:

- ` ` (empty value) or `none` to disable any tracing (default)
- `gc`
- `sched`
- `gc,sched`
- `all` to enable everything

The traces are printed to the standard error by default, but this can be changed at runtime with the `CRYSTAL_TRACE_FILE` environment variable. For example `trace.log`. You can also redirect the standard error to a file (e.g. `2> trace.log` on UNIX shell).

Example tracing calls:

    Crystal.trace :sched, "spawn", fiber: fiber
    Crystal.trace :gc, "malloc", size: size, atomic: 1

**Technical note:** tracing happens before the stdlib is initialized, so the implementation must rely on some `LibC` methods directly (i.e. read environment variable, write to file descriptor) and can't use the core/stdlib abstractions.

Co-authored-by: Johannes Müller <straightshoota@gmail.com>